day-ten/Backend/requirements.txt
Django==3.1.7
daphne==3.0.1
pillow==8.0.1
psycopg2==2.8.5 --no-binary psycopg2
channels==3.0.3
channels-redis==3.2.0
django-channels-jwt-auth-middleware==1.0.0
djoser==2.1.0
djangorestframework_simplejwt==4.6.0
djangorestframework==3.12.2
django-filter==2.4.0
django-cors-headers==3.7.0
drf-yasg==1.20.0
fcm-django==0.3.7
celery==5.1.2
django-redis==5.0.0